refactor: use manual collection in a json
